challenge_two/requirements.txt
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.3
certifi==2019.3.9
chardet==3.0.4
Click==7.0
coverage==4.5.3
coveralls==1.7.0
docopt==0.6.2
Flask==1.0.2
idna==2.8
isort==4.3.16
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==6.0.0
pathlib2==2.3.3
pluggy==0.9.0
py==1.8.0
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.3.1
pytest==4.3.1
pytest-cov==2.6.1
requests==2.21.0
six==1.12.0
typed-ast==1.3.1
urllib3==1.24.1
Werkzeug==0.15.1
wrapt==1.11.1
